+++ includes/elementor/class-llms-elementor-widgets.php
@@ -276,6 +276,9 @@ class LLMS_Elementor_Widgets {
 			new LLMS_Elementor_Widget_Student_Profile(),
 			new LLMS_Elementor_Widget_Student_Dashboard(),
 			new LLMS_Elementor_Widget_Access_Groups(),
+			new LLMS_Elementor_Widget_Materials(),
+			new LLMS_Elementor_Widget_Site_Header(),
+			new LLMS_Elementor_Widget_Site_Footer(),
 		);
 
 		self::$registered_widgets = array();